Bot/qa-bot-llm/requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
dataclasses-json==0.6.3
distro==1.8.0
exceptiongroup==1.2.0
faiss-cpu==1.7.4
fastapi==0.105.0
frozenlist==1.4.1
greenlet==3.0.2
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
jq==1.6.0
jsonpatch==1.33
jsonpointer==2.4
langchain==0.0.352
langchain-community==0.0.5
langchain-core==0.1.2
langserve==0.0.37
langsmith==0.0.72
marshmallow==3.20.1
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.26.2
openai==1.6.0
orjson==3.9.10
packaging==23.2
pydantic==2.5.2
pydantic_core==2.14.5
pypdf==3.17.3
python-dotenv==1.0.0
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
sniffio==1.3.0
SQLAlchemy==2.0.23
sse-starlette==1.8.2
starlette==0.27.0
tenacity==8.2.3
tiktoken==0.5.2
tqdm==4.66.1
typing-inspect==0.9.0
typing_extensions==4.9.0
urllib3==2.1.0
uvicorn==0.25.0
yarl==1.9.4
